+++ tests/mda/test_sequential.py
@@ -22,8 +22,12 @@
 
 from __future__ import division, unicode_literals
 
-import numpy as np
+import sys
 
+import numpy as np
+import pytest
+
+from gemseo.core.parallel_execution import IS_WIN
 from gemseo.mda.jacobi import MDAJacobi
 from gemseo.mda.newton import MDANewtonRaphson
 from gemseo.mda.sequential_mda import GSNewtonMDA, MDASequential
@@ -79,6 +83,10 @@ def test_log_convergence(sellar_disciplines):
         assert not sub_mda.log_convergence
 
 
+@pytest.mark.skipif(
+    sys.version_info < (3, 7) and IS_WIN,
+    reason="Subprocesses in ParallelExecution may hang randomly for Python < 3.7 on Windows.",
+)
 def test_parallel_doe(generate_parallel_doe_data):
     """Test the execution of GaussSeidel in parallel.